Move vault tests to separate file

+++ lib/tests/gc.rs
@@ -224,7 +224,7 @@ async fn write_to_file(file: &mut File, size: usize) {
     file.write(&common::random_content(size)).await.unwrap();
 }
 
-async fn expect_block_count(repo: &Repository, expected_block_count: usize) {
+async fn expect_block_count(repo: &Repository, expected_block_count: u64) {
     common::eventually(repo, || async {
         repo.count_blocks().await.unwrap() == expected_block_count
     })
